refactor: separate components from `useForm` into individual files. | fix: :initial-value prop type
