+++ src/components/index.ts
@@ -19,27 +19,3 @@ export type FormType<T extends Record<string, any> = Record<string, any>> = {
 	reset: () => void;
 	flush: () => void;
 }
-
-export type HTMLInputAttributeType =
-| 'button'
-| 'checkbox'
-| 'color'
-| 'date'
-| 'datetime-local'
-| 'email'
-| 'file'
-| 'hidden'
-| 'image'
-| 'month'
-| 'number'
-| 'password'
-| 'radio'
-| 'range'
-| 'reset'
-| 'search'
-| 'submit'
-| 'tel'
-| 'text'
-| 'time'
-| 'url'
-| 'week';
