+++ src/composable/useForm.ts
@@ -1,8 +1,7 @@
-/* eslint-disable vue/one-component-per-file */
-import { createFormDataFromObject, deleteByPath, EventEmitter, fetcher, flattenObject, getValueByPath, mergeDeep, normalizeChildren, resolveTag } from '@/utils/utils';
-import { computed, defineComponent, h, inject, InputHTMLAttributes, nextTick, onMounted, PropType, provide, ref, resolveDynamicComponent, SlotsType, toValue } from 'vue';
-import { forms } from '@/utils/store';
-import { useField } from './useField';
+import { FieldComp } from '@/components/Field';
+import { FieldArrayComp } from '@/components/FieldArray';
+import { ErrorComp } from '@/components/Error';
+import { FormCompBase, FormOptions } from '@/components/Form';
 
 type GetNestedArray<T> = T extends (infer U)[]
 	? `${GetNestedArray<U>}[]` | `${GetNestedArray<U>}[${number}]`
@@ -22,547 +21,6 @@ export type GetKeys<T extends Record<string, any>> = keyof {
 		T[K] extends object ? `${K & string}.${GetNestedArray<T[K]> & string}` : `${K & string}`
 	)]: any
 }
-
-export type ExtractValue<T, K extends string> =
-	K extends `${infer Root}.${infer Rest}`
-		? Root extends keyof T
-			? ExtractValue<T[Root], Rest>
-			: Root extends `${infer ArrayRoot}[${number}]`
-				? ArrayRoot extends keyof T
-					? T[ArrayRoot] extends (infer U)[]
-						? ExtractValue<U, Rest>
-						: never
-					: never
-				: never
-	: K extends `${infer Root}[${number}]`
-		? Root extends keyof T
-			? T[Root] extends (infer U)
-				? U
-				: never
-			: never
-	: K extends keyof T
-		? T[K]
-		: never;
-
-type RecursivePartial<T> = {
-	[P in keyof T]?: RecursivePartial<T[P]>;
-};
-
-type FormType<T extends Record<string, any>> = {
-	enctype?: 'application/x-www-form-urlencoded' | 'multipart/form-data';
-	validationSchema?: any;
-	action?: string;
-	initialValues?: RecursivePartial<T>;
-	name?: string;
-	preserve?: boolean;
-	onValueChange?: (value?: any) => void;
-	onSubmit?: (value?: any, $event?: SubmitEvent) => void | Promise<any>;
-}
-
-export type FieldType<T extends Record<string, any>> = {
-	[K in GetKeys<T>]: {
-		name: K;
-		default?: ExtractValue<T, K>;
-	} & {
-		error?: any;
-		ignore?: boolean;
-		trueValue?: any;
-		modelValue?: any;
-		falseValue?: any;
-		preserve?: boolean;
-		as?: 'input' | 'select';
-	} & InputHTMLAttributes
-}[GetKeys<T>];
-
-type FieldArrayType<T extends Record<string, any>> = {
-	name: GetKeys<T>;
-	error?: any;
-	ignore?: boolean;
-	initialValues?: any[];
-	preserve?: boolean;
-	default?: any[];
-}
-
-type FormOptions<T extends Record<string, any>> = {
-	initialValues: Partial<T>
-}
-
-const FormCompBase = <T extends Record<string, any> = Record<string, any>>(opt?: FormOptions<T>) => {
-	let uid: number | string = Math.floor(Math.random() * Date.now());
-	let originalForm = Object.create({});
-	let _values = {};
-	const isSubmitting = ref<boolean>(false);
-
-	const handleSubmit = (cb?: (data?: T) => void | Promise<any>) => {
-		return async () => {
-			await cb?.(_values as T);
-		};
-	};
-
-	const setError = (name: GetKeys<T>, error: any) => {
-		if (getValueByPath(forms[uid].values, name as unknown as string)) {
-			getValueByPath(forms[uid].values, name as unknown as string).error = error;
-		}
-	};
-
-	const setInitalValues = (initials: Partial<T>) => {
-		forms[uid].initialValues = mergeDeep(flattenObject(forms[uid].values), initials);
-		forms[uid].key++;
-	};
-
-	const reset = () => {
-		EventEmitter.emit('reset');
-		forms[uid].values = JSON.parse(originalForm);
-		forms[uid].key++;
-	};
-
-	const cmp = defineComponent(
-		(props: FormType<T>, { slots, emit, attrs, expose }) => {
-			props.name && (uid = props.name);
-
-			const updateField = (name: string, value: any) => {
-				if (getValueByPath(forms[uid].values, name)) {
-					getValueByPath(forms[uid].values, name).value = value;
-
-					EventEmitter.emit('value-change', uid);
-				}
-			};
-
-			EventEmitter.on('value-change', (id: string) => {
-				if (id === uid && props?.onValueChange) {
-					emit('value-change', values.value);
-				}
-			});
-
-			const submit = async ($event: any) => {
-				_values = values.value;
-				if (props.validationSchema) {
-					const result = await props.validationSchema.parse(flattenObject(forms[uid].values));
-
-					if (result.errors.length) {
-						result.errors.forEach((err: any) => {
-							setError(err.key, err.message);
-						});
-
-						return $event.preventDefault();
-					}
-				}
-
-				if (props?.enctype === 'multipart/form-data') {
-					_values = createFormDataFromObject(flattenObject(forms[uid].values));
-				}
-
-				if (props.action) {
-					return;
-				}
-				$event.preventDefault();
-				isSubmitting.value = true;
-				await fetcher(props.onSubmit?.(_values, $event));
-				isSubmitting.value = false;
-			};
-
-			const flush = () => {
-				forms[uid].initialValues = Object.create({});
-				forms[uid].values = Object.create({});
-				forms[uid].key++;
-			};
-
-			/*---------------------------------------------
-			/  COMPUTED
-			---------------------------------------------*/
-			const errors = computed(() => {
-				return flattenObject(forms[uid].values, 'error');
-			});
-			const values = computed(() => {
-				return flattenObject(forms[uid].values) as T;
-			});
-
-			/*---------------------------------------------
-			/  CREATED
-			---------------------------------------------*/
-			if (!forms[uid]) {
-				forms[uid] = {
-					values: Object.create({}),
-					initialValues: opt?.initialValues || props.initialValues || {},
-					key: 0,
-				};
-			} else {
-				props.initialValues && (forms[uid].initialValues = opt?.initialValues || props.initialValues);
-			}
-
-			if (props.validationSchema && typeof props.validationSchema.cast === 'function') {
-				forms[uid].initialValues = props.initialValues ? mergeDeep(props.initialValues, opt?.initialValues || {}, props.validationSchema.cast(flattenObject(forms[uid].values))) : mergeDeep(opt?.initialValues || {}, props.validationSchema.cast(flattenObject(forms[uid].values)));
-			}
-
-			/*---------------------------------------------
-			/  HOOKS
-			---------------------------------------------*/
-			onMounted(() => {
-				originalForm = JSON.stringify(forms[uid].values);
-			});
-
-			provide('formData', {
-				uid,
-				preserve: props.preserve,
-			});
-
-			expose({
-				values,
-				errors,
-				setError,
-				updateField,
-				reset,
-				flush,
-				isSubmitting,
-			});
-
-
-			return () => {
-				return h('form',
-					{ ...props, ...attrs, ...emit, key: forms[uid].key, onSubmit: submit },
-					slots.default?.({ values: values.value, errors: errors.value }),
-				);
-			};
-		},
-		{
-			name: 'FormifyForm',
-			props: {
-				enctype: {
-					type: String as PropType<'application/x-www-form-urlencoded' | 'multipart/form-data'>,
-					default: undefined,
-				},
-				validationSchema: {
-					type: Object as PropType<Record<string, any>>,
-					default: undefined,
-				},
-				initialValues: {
-					type: Object as PropType<any>,
-					default: undefined,
-				},
-				name: {
-					type: String as PropType<string>,
-					default: undefined,
-				},
-				action: {
-					type: String as PropType<string>,
-					default: undefined,
-				},
-				preserve: {
-					type: Boolean as PropType<boolean>,
-					default: undefined,
-				},
-				onValueChange: {
-					type: Function as PropType<(value?: any) => void>,
-					default: undefined,
-				},
-				onSubmit: {
-					type: Function as PropType<(value?: any, $event?: SubmitEvent) => void | Promise<any>>,
-					default: undefined,
-				},
-			},
-			slots: Object as SlotsType<{
-				default: { values: T, errors: any }
-			}>,
-			emits: ['submit', 'value-change'],
-		},
-	);
-
-	return {
-		cmp,
-		handleSubmit,
-		setError,
-		setInitalValues,
-		reset,
-		isSubmitting,
-	};
-};
-
-//@ts-expect-error Object props throw error with `FieldWithAutoComplete` but it's working actually
-const FieldComp = <T extends Record<string, any> = Record<string, any>>() => defineComponent<
-	FieldType<T>,
-	any,
-	string,
-	SlotsType<{
-		default: { field: { value: any }, error: any }
-	}>
->(
-	(props: FieldType<T>, { emit, slots, attrs: baseAttrs }) => {
-		const {
-			value,
-			onInput,
-			onFocus,
-			getError,
-		} = useField(props, emit);
-
-		const sharedProps = computed(() => {
-			const attrs: Record<string, any> = {
-				...baseAttrs,
-				name: props.name,
-				error: getError(),
-				onInput: (evt: any) => {
-					onInput(evt);
-					if (typeof baseAttrs.onInput === 'function') {
-						baseAttrs.onInput();
-					}
-				},
-				onChange: (evt: any) => {
-					onInput(evt);
-					if (typeof baseAttrs.onChange === 'function') {
-						baseAttrs.onChange();
-					}
-				},
-				onFocus: () => {
-					onFocus();
-					if (typeof baseAttrs.onFocus === 'function') {
-						baseAttrs.onFocus();
-					}
-				},
-				onBlur: () => {
-					if (typeof baseAttrs.onBlur === 'function') {
-						baseAttrs.onBlur();
-					}
-				},
-			};
-
-			if (attrs.type === 'checkbox' && value.value) {
-				attrs.checked = true;
-			}
-
-			if (baseAttrs.type !== 'file' && (props.as !== 'select' && !attrs.multiple)) {
-				attrs.value = value.value;
-			}
-
-			return attrs;
-		});
-
-		const slotProps = () => {
-			return {
-				field: {
-					...sharedProps.value,
-					modelValue: value.value,
-				},
-				componentField: {
-					value: value.value,
-				},
-				value: value.value,
-				modelValue: value.value,
-				error: getError(),
-			};
-		};
-
-		return () => {
-			const tag = resolveDynamicComponent(resolveTag(props, slots)) as string;
-			const children = normalizeChildren(tag, slots, slotProps);
-			if (tag) {
-				return h(tag,
-					{
-						...props,
-						...sharedProps.value,
-					},
-					children,
-				);
-			} else {
-				return children;
-			}
-		};
-	},
-	{
-		name: 'Field',
-		inheritAttrs: false,
-		props: {
-			name: {
-				type: String as PropType<any>,
-				required: true,
-			},
-			default: {
-				type: [String, Array, Boolean, Number, Object] as PropType<any>,
-				default: undefined,
-			},
-			error: {
-				type: String,
-				default: undefined,
-			},
-			ignore: {
-				type: Boolean,
-				default: false,
-			},
-			preserve: {
-				type: Boolean,
-				default: false,
-			},
-			trueValue: {
-				type: [Boolean, String, Number],
-				default: true,
-			},
-			falseValue: {
-				type: [Boolean, String, Number],
-				default: false,
-			},
-			modelValue: {
-				type: [String, Array, Boolean, Number, Object] as PropType<any>,
-				default: undefined,
-			},
-			as: {
-				type: String as PropType<'input' | 'select' | undefined>,
-				default: undefined,
-			},
-		},
-		emits: ['update:modelValue'],
-	},
-);
-
-const FieldArrayComp = <T extends Record<string, any> = Record<string, any>>() => defineComponent(
-	(props: FieldArrayType<T>, { slots, emit, attrs }) => {
-		/*---------------------------------------------
-		/  VARIABLES
-		---------------------------------------------*/
-		const fields = ref<any[]>([]);
-		const { setArrayValue, getError } = useField(props, emit, true);
-		const { uid } = inject('formData', Object.create({}));
-
-		/*---------------------------------------------
-		/  METHODS
-		---------------------------------------------*/
-		const add = () => {
-			!props.ignore && (getValueByPath(forms[uid].values, props.name as string).error = undefined);
-
-			fields.value.push({
-				id: fields.value.length,
-				name: props.name,
-			});
-		};
-
-		const remove = (idx: number) => {
-			getValueByPath(forms[uid].values, props.name as string).error = undefined;
-
-			const removedIndex = fields.value.findIndex(field => field.id === idx);
-			for (let index = removedIndex + 1; index < fields.value.length; index++) {
-				const tmp = JSON.parse(JSON.stringify(getValueByPath(forms[uid].values, `${props.name as string}[${index}]`)));
-				if (getValueByPath(forms[uid].values, `${props.name as string}[${index - 1}]`)) {
-					setArrayValue({
-						[`[${index - 1}]`]: tmp,
-					});
-				}
-			}
-			deleteByPath(forms[uid].values, `${fields.value[fields.value.length - 1].name}[${fields.value.length - 1}]`);
-			fields.value.splice(-1);
-		};
-
-		const init = () => {
-			fields.value = [];
-			const initials = forms[uid].initialValues?.[props.name] || props.initialValues;
-
-			if (initials) {
-				initials.forEach(() => {
-					fields.value.push({
-						id: fields.value.length,
-						name: props.name,
-					});
-				});
-
-				nextTick(() => {
-					if (initials) {
-						initials.forEach((value: any, idx: any) => {
-							if (typeof value === 'object') {
-								Object.keys(value).forEach((key) => {
-									setArrayValue({
-										[`[${idx}]`]: {
-											[key]: {
-												value: value[key],
-											},
-										},
-									});
-								});
-							} else {
-								setArrayValue({
-									[`[${idx}]`]: {
-										value,
-									},
-								});
-							}
-						});
-					}
-				});
-			}
-		};
-		/*---------------------------------------------
-		/  CREATED
-		---------------------------------------------*/
-		init();
-		EventEmitter.on('reset', () => {
-			init();
-		});
-
-		return () => {
-			return h('div',
-				{ ...props, ...attrs, ...emit },
-				slots.default?.({ fields: toValue(fields), add, remove, error: getError() }),
-			);
-		};
-	},
-	{
-		name: 'FieldArray',
-		props: {
-			name: {
-				type: String as PropType<any>,
-				required: true,
-			},
-			error: {
-				type: String as PropType<string>,
-				default: undefined,
-			},
-			ignore: {
-				type: Boolean as PropType<boolean>,
-				default: undefined,
-			},
-			initialValues: {
-				type: Object as PropType<any>,
-				default: undefined,
-			},
-			default: {
-				type: Array as PropType<any[]>,
-				default: undefined,
-			},
-			preserve: {
-				type: Boolean as PropType<boolean>,
-				default: undefined,
-			},
-		},
-		slots: Object as SlotsType<{
-			default: { fields: any[], add: () => void, remove: (idx: number) => void, error: any }
-		}>,
-		emits: ['update:modelValue'],
-	},
-);
-
-const ErrorComp = <T extends Record<string, any> = Record<string, any>>() => defineComponent(
-	(props: { errorFor: GetKeys<T> }, { slots, emit, attrs }) => {
-		const { uid } = inject('formData', Object.create({}));
-
-		const getError = () => {
-			return getValueByPath(forms[uid].values, props.errorFor as string)?.error;
-		};
-
-		return () => {
-			return h('span',
-				{ ...props, ...attrs, ...emit },
-				slots.default ? slots.default?.({ error: getError() }) as SlotsType<{ error: any }> : getError(),
-			);
-		};
-	},
-	{
-		name: 'FieldError',
-		props: {
-			errorFor: {
-				type: String as PropType<any>,
-				default: undefined,
-			},
-		},
-		slots: Object as SlotsType<{
-			default: { error: any }
-		}>,
-	},
-);
 
 export const Form = FormCompBase().cmp;
 export const Field = FieldComp();
